feat(recipe)!: Migrate from sqlite to PostgreSQL
- adjust sqlc settings
- migrate SQL to postgres
- regenerate sql code
- adjust some types
- remove foreign key constraints that are now enforced by postgres but will always be violated bc. auth does not store stuff in here
- add development postgres via docker compose

BREAKING CHANGE: The new PG_CONN_STRING env variable must be set, the service requires a working SQL connection

+++ recipe/recipes/model/model.go
@@ -51,8 +51,8 @@ func RecipeFromDatabaseModel(recipe db.Recipe, ingredients []db.Ingredient) Reci
 		Name:         recipe.RecipeName,
 		Ingredients:  fun.Map(ingredients, IngredientFromDatabaseModel),
 		Directions:   []string{recipe.Directions},
-		TimeEstimate: int(recipe.TimeEstimate.Int64),
+		TimeEstimate: int(recipe.TimeEstimate.Int32),
 		Difficulty:   recipe.Difficulty.String,
-		FeedsPeople:  int(recipe.FeedsPeople.Int64),
+		FeedsPeople:  int(recipe.FeedsPeople.Int32),
 	}
 }
